Moving GroupItemCheckStateChangedEvent to extensions
