+++ app/src/main/java/com/fibelatti/raffler/views/adapters/GroupAdapter.java
@@ -11,8 +11,8 @@ import android.widget.TextView;
 
 import com.fibelatti.raffler.R;
 import com.fibelatti.raffler.models.GroupItem;
+import com.fibelatti.raffler.views.extensions.GroupItemCheckStateChangedEvent;
 import com.fibelatti.raffler.views.utils.BusHelper;
-import com.fibelatti.raffler.views.utils.GroupItemCheckStateChangedEvent;
 import com.squareup.otto.Subscribe;
 
 import java.util.HashSet;
